4.21*: bloglist expansion, step9
modify error

+++ part4/controllers/blogs.js
@@ -49,6 +49,11 @@ blogsRouter.get('/', async (request, response) => {
       if (!decodedToken.id) {
         return res.status(401).json({ error: 'token missing or invalid' })
       }
+
+      const blog = await Blog.findById(req.params.id)
+      if ( blog.user.toString() !== decodedToken.id.toString() ){
+        return res.status(404).json({error: 'permission denied'})
+      }
       const deletedBlog = await Blog.findByIdAndDelete(req.params.id)
       if (!deletedBlog) {
         return res.status(404).json({error: 'blog not found'})
